+++ BTCPayServer.Plugins.LNbank/Services/LightningInvoicePayRequest.cs
@@ -1,4 +1,7 @@
 using System.ComponentModel.DataAnnotations;
+using BTCPayServer.Client.JsonConverters;
+using BTCPayServer.Lightning;
+using Newtonsoft.Json;
 
 namespace BTCPayServer.Plugins.LNbank.Services;
 
@@ -6,5 +9,9 @@ public class LightningInvoicePayRequest
 {
     [Required]
     public string PaymentRequest { get; set; }
+    
     public float? MaxFeePercent { get; set; }
+    
+    [JsonConverter(typeof(LightMoneyJsonConverter))]
+    public LightMoney Amount { get; set; }
 }